+++ backend/app.js
@@ -12,7 +12,7 @@ const { notFound, errorHandler } = require('./middlewares');
 // Routes
 const authRouter = require('./routes/authRoute');
 const userRouter = require('./routes/userRoute');
-
+const contesRouter = require('./routes/contesRoute');
 // Middlewares
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
@@ -21,7 +21,7 @@ app.use(express.urlencoded({ extended: true }));
 app.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerConfigurations));
 app.use('/auth', authRouter);
 app.use('/user', userRouter);
-
+app.use('/contes',contesRouter);
 app.get('/', (req, res) => {
   res.status(200).json({
     message: 'Hello World'
